Better fix instead of `clearEmptyCategories`, check comment:
`this.parameters` holds all parameters, including categories definitions (parameters), this category-parameters have itself a `category_id` property, we was consider it as any regular parameter that we want to identify it's category, and most of the time this `category_id` for this categories-parameters, is null. so whenever a new category(which is a parameter after all) is created, it will consider it as an empty category that should exist and it's associated with this parameter. an "unnamed category" will last forever unless we will make each category has itself a predefined category.